+++ index.py
@@ -1,10 +1,12 @@
 import logging
 import os
+from logging.handlers import RotatingFileHandler
+
+from celery import Celery
 from flask import Flask
 from flask_cors import CORS
 from flask_restful import Api
 from flask_sqlalchemy import SQLAlchemy
-from logging.handlers import RotatingFileHandler
 
 levels = {"DEBUG": logging.DEBUG,
           "INFO": logging.INFO,
@@ -32,6 +34,23 @@ def create_app():
     return app
 
 
+def make_celery(app):
+    celery = Celery(
+        app.import_name,
+        # backend=app.config['CELERY_BACKEND'],
+        # broker=app.config['CELERY_BROKER_URL']
+    )
+    celery.conf.update(app.config)
+
+    class ContextTask(celery.Task):
+        def __call__(self, *args, **kwargs):
+            with app.app_context():
+                return self.run(*args, **kwargs)
+
+    celery.Task = ContextTask
+    return celery
+
+
 basedir = os.path.abspath(os.path.dirname(__file__))
 static_folder = basedir + '/acciom_ui/build/'
 app = create_app()
@@ -39,3 +58,5 @@ CORS(app)
 app.url_map.strict_slashes = False
 db = SQLAlchemy(app)
 api = Api(app)
+
+celery = make_celery(app)
